Rename WASM libraries to `cosmian_*`
