+++ build_wasm_for_nodejs.sh
@@ -24,9 +24,9 @@ build_wasm_bindgen() {
 }
 
 rm -rf "${WASM_DIR}"
-build_wasm_bindgen git@github.com:Cosmian/abe_gpsw.git abe_gpsw abe/gpsw
-build_wasm_bindgen git@github.com:Cosmian/cover_crypt.git cover_crypt abe/cover_crypt
-build_wasm_bindgen git@gitlab.cosmian.com:core/findex.git findex findex
+build_wasm_bindgen git@github.com:Cosmian/abe_gpsw.git abe_gpsw cosmian_abe_gpsw
+build_wasm_bindgen git@github.com:Cosmian/cover_crypt.git cover_crypt cosmian_cover_crypt
+build_wasm_bindgen git@gitlab.cosmian.com:core/findex.git findex cosmian_findex
 
 cd "${CUR_DIR}"
 npm install
