+++ src/crypto/abe/core/hybrid_crypto/cover_crypt/decryption.ts
@@ -3,10 +3,10 @@ import {
   webassembly_decrypt_hybrid_block,
   webassembly_decrypt_hybrid_header,
   webassembly_get_encrypted_header_size,
-} from "cover_crypt";
-import { logger } from "utils/logger";
+} from "cosmian_cover_crypt";
 import { ClearTextHeader } from "crypto/abe/interfaces/cleartext_header";
 import { HybridDecryption } from "crypto/abe/interfaces/decryption";
+import { logger } from "utils/logger";
 
 /**
  * This class exposes the ABE primitives.
