+++ src/crypto/abe/core/hybrid_crypto/gpsw/encryption.ts
@@ -4,15 +4,15 @@ import {
   webassembly_encrypt_hybrid_block,
   webassembly_encrypt_hybrid_header,
   webassembly_encrypt_hybrid_header_using_cache,
-} from "abe_gpsw";
+} from "cosmian_abe_gpsw";
+import { HybridEncryption } from "crypto/abe/interfaces/encryption";
 import { logger } from "utils/logger";
+import { fromBeBytes } from "utils/utils";
 import { EncryptedHeader } from "../../../interfaces/encrypted_header";
 import {
   AbeEncryptionParameters,
   Metadata,
 } from "../../../interfaces/encryption_parameters";
-import { HybridEncryption } from "crypto/abe/interfaces/encryption";
-import { fromBeBytes } from "utils/utils";
 
 /**
  * This class exposes the ABE primitives.
