+++ tests/wasm_lib/cosmian_abe_gpsw/cosmian_abe_gpsw.d.ts
