+++ tests/wasm_lib/cosmian_abe_gpsw/cosmian_abe_gpsw_bg.wasm.d.ts
@@ -16,7 +16,7 @@ export function webassembly_destroy_encryption_cache(a: number, b: number): void
 export function webassembly_encrypt_hybrid_header_using_cache(a: number, b: number, c: number, d: number, e: number): void;
 export function webassembly_encrypt_hybrid_block(a: number, b: number, c: number, d: number, e: number, f: number): void;
 export function __wbindgen_add_to_stack_pointer(a: number): number;
-export function __wbindgen_malloc(a: number): number;
-export function __wbindgen_realloc(a: number, b: number, c: number): number;
-export function __wbindgen_free(a: number, b: number): void;
-export function __wbindgen_exn_store(a: number): void;
+export function __wbindgen_export_0(a: number): number;
+export function __wbindgen_export_1(a: number, b: number, c: number): number;
+export function __wbindgen_export_2(a: number, b: number): void;
+export function __wbindgen_export_3(a: number): void;
