+++ tests/wasm_lib/cosmian_cover_crypt/cosmian_cover_crypt.d.ts
