+++ tests/wasm_lib/cosmian_cover_crypt/cosmian_cover_crypt.js
@@ -1,28 +1,9 @@
-import * as wasm from './cover_crypt_bg.wasm';
+let imports = {};
+imports['__wbindgen_placeholder__'] = module.exports;
+let wasm;
+const { TextDecoder, TextEncoder } = require(`util`);
 
-const heap = new Array(32).fill(undefined);
-
-heap.push(undefined, null, true, false);
-
-function getObject(idx) { return heap[idx]; }
-
-let heap_next = heap.length;
-
-function dropObject(idx) {
-    if (idx < 36) return;
-    heap[idx] = heap_next;
-    heap_next = idx;
-}
-
-function takeObject(idx) {
-    const ret = getObject(idx);
-    dropObject(idx);
-    return ret;
-}
-
-const lTextDecoder = typeof TextDecoder === 'undefined' ? (0, module.require)('util').TextDecoder : TextDecoder;
-
-let cachedTextDecoder = new lTextDecoder('utf-8', { ignoreBOM: true, fatal: true });
+let cachedTextDecoder = new TextDecoder('utf-8', { ignoreBOM: true, fatal: true });
 
 cachedTextDecoder.decode();
 
@@ -39,6 +20,12 @@ function getStringFromWasm0(ptr, len) {
     return cachedTextDecoder.decode(getUint8Memory0().subarray(ptr, ptr + len));
 }
 
+const heap = new Array(32).fill(undefined);
+
+heap.push(undefined, null, true, false);
+
+let heap_next = heap.length;
+
 function addHeapObject(obj) {
     if (heap_next === heap.length) heap.push(heap.length + 1);
     const idx = heap_next;
@@ -46,6 +33,20 @@ function addHeapObject(obj) {
 
     heap[idx] = obj;
     return idx;
+}
+
+function getObject(idx) { return heap[idx]; }
+
+function dropObject(idx) {
+    if (idx < 36) return;
+    heap[idx] = heap_next;
+    heap_next = idx;
+}
+
+function takeObject(idx) {
+    const ret = getObject(idx);
+    dropObject(idx);
+    return ret;
 }
 
 let cachedInt32Memory0 = new Int32Array();
@@ -64,7 +65,7 @@ function getInt32Memory0() {
 * @param {Uint8Array} policy_bytes
 * @returns {Uint8Array}
 */
-export function webassembly_generate_master_keys(policy_bytes) {
+module.exports.webassembly_generate_master_keys = function(policy_bytes) {
     try {
         const retptr = wasm.__wbindgen_add_to_stack_pointer(-16);
         wasm.webassembly_generate_master_keys(retptr, addHeapObject(policy_bytes));
@@ -78,13 +79,11 @@ export function webassembly_generate_master_keys(policy_bytes) {
     } finally {
         wasm.__wbindgen_add_to_stack_pointer(16);
     }
-}
+};
 
 let WASM_VECTOR_LEN = 0;
 
-const lTextEncoder = typeof TextEncoder === 'undefined' ? (0, module.require)('util').TextEncoder : TextEncoder;
-
-let cachedTextEncoder = new lTextEncoder('utf-8');
+let cachedTextEncoder = new TextEncoder('utf-8');
 
 const encodeString = (typeof cachedTextEncoder.encodeInto === 'function'
     ? function (arg, view) {
@@ -148,10 +147,10 @@ function passStringToWasm0(arg, malloc, realloc) {
 * @param {Uint8Array} policy_bytes
 * @returns {Uint8Array}
 */
-export function webassembly_generate_user_private_key(master_private_key_bytes, access_policy_str, policy_bytes) {
+module.exports.webassembly_generate_user_private_key = function(master_private_key_bytes, access_policy_str, policy_bytes) {
     try {
         const retptr = wasm.__wbindgen_add_to_stack_pointer(-16);
-        const ptr0 = passStringToWasm0(access_policy_str, wasm.__wbindgen_malloc, wasm.__wbindgen_realloc);
+        const ptr0 = passStringToWasm0(access_policy_str, wasm.__wbindgen_export_0, wasm.__wbindgen_export_1);
         const len0 = WASM_VECTOR_LEN;
         wasm.webassembly_generate_user_private_key(retptr, addHeapObject(master_private_key_bytes), ptr0, len0, addHeapObject(policy_bytes));
         var r0 = getInt32Memory0()[retptr / 4 + 0];
@@ -164,7 +163,7 @@ export function webassembly_generate_user_private_key(master_private_key_bytes, 
     } finally {
         wasm.__wbindgen_add_to_stack_pointer(16);
     }
-}
+};
 
 /**
 * Rotate attributes, changing their underlying values with that of an unused
@@ -176,7 +175,7 @@ export function webassembly_generate_user_private_key(master_private_key_bytes, 
 * @param {Uint8Array} policy_bytes
 * @returns {string}
 */
-export function webassembly_rotate_attributes(attributes_bytes, policy_bytes) {
+module.exports.webassembly_rotate_attributes = function(attributes_bytes, policy_bytes) {
     try {
         const retptr = wasm.__wbindgen_add_to_stack_pointer(-16);
         wasm.webassembly_rotate_attributes(retptr, addHeapObject(attributes_bytes), addHeapObject(policy_bytes));
@@ -193,16 +192,16 @@ export function webassembly_rotate_attributes(attributes_bytes, policy_bytes) {
         return getStringFromWasm0(ptr0, len0);
     } finally {
         wasm.__wbindgen_add_to_stack_pointer(16);
-        wasm.__wbindgen_free(ptr0, len0);
+        wasm.__wbindgen_export_2(ptr0, len0);
     }
-}
+};
 
 /**
 * Extract header from encrypted bytes
 * @param {Uint8Array} encrypted_bytes
 * @returns {number}
 */
-export function webassembly_get_encrypted_header_size(encrypted_bytes) {
+module.exports.webassembly_get_encrypted_header_size = function(encrypted_bytes) {
     try {
         const retptr = wasm.__wbindgen_add_to_stack_pointer(-16);
         wasm.webassembly_get_encrypted_header_size(retptr, addHeapObject(encrypted_bytes));
@@ -216,7 +215,7 @@ export function webassembly_get_encrypted_header_size(encrypted_bytes) {
     } finally {
         wasm.__wbindgen_add_to_stack_pointer(16);
     }
-}
+};
 
 /**
 * @param {Uint8Array} metadata_bytes
@@ -225,7 +224,7 @@ export function webassembly_get_encrypted_header_size(encrypted_bytes) {
 * @param {Uint8Array} public_key_bytes
 * @returns {Uint8Array}
 */
-export function webassembly_encrypt_hybrid_header(metadata_bytes, policy_bytes, attributes_bytes, public_key_bytes) {
+module.exports.webassembly_encrypt_hybrid_header = function(metadata_bytes, policy_bytes, attributes_bytes, public_key_bytes) {
     try {
         const retptr = wasm.__wbindgen_add_to_stack_pointer(-16);
         wasm.webassembly_encrypt_hybrid_header(retptr, addHeapObject(metadata_bytes), addHeapObject(policy_bytes), addHeapObject(attributes_bytes), addHeapObject(public_key_bytes));
@@ -239,7 +238,7 @@ export function webassembly_encrypt_hybrid_header(metadata_bytes, policy_bytes, 
     } finally {
         wasm.__wbindgen_add_to_stack_pointer(16);
     }
-}
+};
 
 /**
 * Decrypt with a user decryption key an encrypted header
@@ -248,7 +247,7 @@ export function webassembly_encrypt_hybrid_header(metadata_bytes, policy_bytes, 
 * @param {Uint8Array} encrypted_header_bytes
 * @returns {Uint8Array}
 */
-export function webassembly_decrypt_hybrid_header(user_decryption_key_bytes, encrypted_header_bytes) {
+module.exports.webassembly_decrypt_hybrid_header = function(user_decryption_key_bytes, encrypted_header_bytes) {
     try {
         const retptr = wasm.__wbindgen_add_to_stack_pointer(-16);
         wasm.webassembly_decrypt_hybrid_header(retptr, addHeapObject(user_decryption_key_bytes), addHeapObject(encrypted_header_bytes));
@@ -262,7 +261,7 @@ export function webassembly_decrypt_hybrid_header(user_decryption_key_bytes, enc
     } finally {
         wasm.__wbindgen_add_to_stack_pointer(16);
     }
-}
+};
 
 function isLikeNone(x) {
     return x === undefined || x === null;
@@ -275,7 +274,7 @@ function isLikeNone(x) {
 * @param {Uint8Array} plaintext_bytes
 * @returns {Uint8Array}
 */
-export function webassembly_encrypt_hybrid_block(symmetric_key_bytes, uid_bytes, block_number, plaintext_bytes) {
+module.exports.webassembly_encrypt_hybrid_block = function(symmetric_key_bytes, uid_bytes, block_number, plaintext_bytes) {
     try {
         const retptr = wasm.__wbindgen_add_to_stack_pointer(-16);
         wasm.webassembly_encrypt_hybrid_block(retptr, addHeapObject(symmetric_key_bytes), isLikeNone(uid_bytes) ? 0 : addHeapObject(uid_bytes), !isLikeNone(block_number), isLikeNone(block_number) ? 0 : block_number, addHeapObject(plaintext_bytes));
@@ -289,7 +288,7 @@ export function webassembly_encrypt_hybrid_block(symmetric_key_bytes, uid_bytes,
     } finally {
         wasm.__wbindgen_add_to_stack_pointer(16);
     }
-}
+};
 
 /**
 * Symmetrically Decrypt encrypted data in a block.
@@ -299,7 +298,7 @@ export function webassembly_encrypt_hybrid_block(symmetric_key_bytes, uid_bytes,
 * @param {Uint8Array} encrypted_bytes
 * @returns {Uint8Array}
 */
-export function webassembly_decrypt_hybrid_block(symmetric_key_bytes, uid_bytes, block_number, encrypted_bytes) {
+module.exports.webassembly_decrypt_hybrid_block = function(symmetric_key_bytes, uid_bytes, block_number, encrypted_bytes) {
     try {
         const retptr = wasm.__wbindgen_add_to_stack_pointer(-16);
         wasm.webassembly_decrypt_hybrid_block(retptr, addHeapObject(symmetric_key_bytes), isLikeNone(uid_bytes) ? 0 : addHeapObject(uid_bytes), !isLikeNone(block_number), isLikeNone(block_number) ? 0 : block_number, addHeapObject(encrypted_bytes));
@@ -313,13 +312,13 @@ export function webassembly_decrypt_hybrid_block(symmetric_key_bytes, uid_bytes,
     } finally {
         wasm.__wbindgen_add_to_stack_pointer(16);
     }
-}
+};
 
 function handleError(f, args) {
     try {
         return f.apply(this, args);
     } catch (e) {
-        wasm.__wbindgen_exn_store(addHeapObject(e));
+        wasm.__wbindgen_export_3(addHeapObject(e));
     }
 }
 
@@ -327,149 +326,157 @@ function getArrayU8FromWasm0(ptr, len) {
     return getUint8Memory0().subarray(ptr / 1, ptr / 1 + len);
 }
 
-export function __wbindgen_object_drop_ref(arg0) {
-    takeObject(arg0);
-};
-
-export function __wbindgen_string_new(arg0, arg1) {
+module.exports.__wbindgen_string_new = function(arg0, arg1) {
     const ret = getStringFromWasm0(arg0, arg1);
     return addHeapObject(ret);
 };
 
-export function __wbg_randomFillSync_91e2b39becca6147() { return handleError(function (arg0, arg1, arg2) {
+module.exports.__wbindgen_object_drop_ref = function(arg0) {
+    takeObject(arg0);
+};
+
+module.exports.__wbg_randomFillSync_91e2b39becca6147 = function() { return handleError(function (arg0, arg1, arg2) {
     getObject(arg0).randomFillSync(getArrayU8FromWasm0(arg1, arg2));
 }, arguments) };
 
-export function __wbg_getRandomValues_b14734aa289bc356() { return handleError(function (arg0, arg1) {
+module.exports.__wbg_getRandomValues_b14734aa289bc356 = function() { return handleError(function (arg0, arg1) {
     getObject(arg0).getRandomValues(getObject(arg1));
 }, arguments) };
 
-export function __wbg_process_e56fd54cf6319b6c(arg0) {
+module.exports.__wbg_process_e56fd54cf6319b6c = function(arg0) {
     const ret = getObject(arg0).process;
     return addHeapObject(ret);
 };
 
-export function __wbindgen_is_object(arg0) {
+module.exports.__wbindgen_is_object = function(arg0) {
     const val = getObject(arg0);
     const ret = typeof(val) === 'object' && val !== null;
     return ret;
 };
 
-export function __wbg_versions_77e21455908dad33(arg0) {
+module.exports.__wbg_versions_77e21455908dad33 = function(arg0) {
     const ret = getObject(arg0).versions;
     return addHeapObject(ret);
 };
 
-export function __wbg_node_0dd25d832e4785d5(arg0) {
+module.exports.__wbg_node_0dd25d832e4785d5 = function(arg0) {
     const ret = getObject(arg0).node;
     return addHeapObject(ret);
 };
 
-export function __wbindgen_is_string(arg0) {
+module.exports.__wbindgen_is_string = function(arg0) {
     const ret = typeof(getObject(arg0)) === 'string';
     return ret;
 };
 
-export function __wbg_crypto_b95d7173266618a9(arg0) {
+module.exports.__wbg_crypto_b95d7173266618a9 = function(arg0) {
     const ret = getObject(arg0).crypto;
     return addHeapObject(ret);
 };
 
-export function __wbg_msCrypto_5a86d77a66230f81(arg0) {
+module.exports.__wbg_msCrypto_5a86d77a66230f81 = function(arg0) {
     const ret = getObject(arg0).msCrypto;
     return addHeapObject(ret);
 };
 
-export function __wbg_static_accessor_NODE_MODULE_26b231378c1be7dd() {
+module.exports.__wbg_static_accessor_NODE_MODULE_26b231378c1be7dd = function() {
     const ret = module;
     return addHeapObject(ret);
 };
 
-export function __wbg_require_0db1598d9ccecb30() { return handleError(function (arg0, arg1, arg2) {
+module.exports.__wbg_require_0db1598d9ccecb30 = function() { return handleError(function (arg0, arg1, arg2) {
     const ret = getObject(arg0).require(getStringFromWasm0(arg1, arg2));
     return addHeapObject(ret);
 }, arguments) };
 
-export function __wbg_newnoargs_971e9a5abe185139(arg0, arg1) {
+module.exports.__wbg_newnoargs_971e9a5abe185139 = function(arg0, arg1) {
     const ret = new Function(getStringFromWasm0(arg0, arg1));
     return addHeapObject(ret);
 };
 
-export function __wbg_call_33d7bcddbbfa394a() { return handleError(function (arg0, arg1) {
+module.exports.__wbg_call_33d7bcddbbfa394a = function() { return handleError(function (arg0, arg1) {
     const ret = getObject(arg0).call(getObject(arg1));
     return addHeapObject(ret);
 }, arguments) };
 
-export function __wbindgen_object_clone_ref(arg0) {
+module.exports.__wbindgen_object_clone_ref = function(arg0) {
     const ret = getObject(arg0);
     return addHeapObject(ret);
 };
 
-export function __wbg_self_fd00a1ef86d1b2ed() { return handleError(function () {
+module.exports.__wbg_self_fd00a1ef86d1b2ed = function() { return handleError(function () {
     const ret = self.self;
     return addHeapObject(ret);
 }, arguments) };
 
-export function __wbg_window_6f6e346d8bbd61d7() { return handleError(function () {
+module.exports.__wbg_window_6f6e346d8bbd61d7 = function() { return handleError(function () {
     const ret = window.window;
     return addHeapObject(ret);
 }, arguments) };
 
-export function __wbg_globalThis_3348936ac49df00a() { return handleError(function () {
+module.exports.__wbg_globalThis_3348936ac49df00a = function() { return handleError(function () {
     const ret = globalThis.globalThis;
     return addHeapObject(ret);
 }, arguments) };
 
-export function __wbg_global_67175caf56f55ca9() { return handleError(function () {
+module.exports.__wbg_global_67175caf56f55ca9 = function() { return handleError(function () {
     const ret = global.global;
     return addHeapObject(ret);
 }, arguments) };
 
-export function __wbindgen_is_undefined(arg0) {
+module.exports.__wbindgen_is_undefined = function(arg0) {
     const ret = getObject(arg0) === undefined;
     return ret;
 };
 
-export function __wbg_buffer_34f5ec9f8a838ba0(arg0) {
+module.exports.__wbg_buffer_34f5ec9f8a838ba0 = function(arg0) {
     const ret = getObject(arg0).buffer;
     return addHeapObject(ret);
 };
 
-export function __wbg_newwithbyteoffsetandlength_88fdad741db1b182(arg0, arg1, arg2) {
+module.exports.__wbg_newwithbyteoffsetandlength_88fdad741db1b182 = function(arg0, arg1, arg2) {
     const ret = new Uint8Array(getObject(arg0), arg1 >>> 0, arg2 >>> 0);
     return addHeapObject(ret);
 };
 
-export function __wbg_new_cda198d9dbc6d7ea(arg0) {
+module.exports.__wbg_new_cda198d9dbc6d7ea = function(arg0) {
     const ret = new Uint8Array(getObject(arg0));
     return addHeapObject(ret);
 };
 
-export function __wbg_set_1a930cfcda1a8067(arg0, arg1, arg2) {
+module.exports.__wbg_set_1a930cfcda1a8067 = function(arg0, arg1, arg2) {
     getObject(arg0).set(getObject(arg1), arg2 >>> 0);
 };
 
-export function __wbg_length_51f19f73d6d9eff3(arg0) {
+module.exports.__wbg_length_51f19f73d6d9eff3 = function(arg0) {
     const ret = getObject(arg0).length;
     return ret;
 };
 
-export function __wbg_newwithlength_66e5530e7079ea1b(arg0) {
+module.exports.__wbg_newwithlength_66e5530e7079ea1b = function(arg0) {
     const ret = new Uint8Array(arg0 >>> 0);
     return addHeapObject(ret);
 };
 
-export function __wbg_subarray_270ff8dd5582c1ac(arg0, arg1, arg2) {
+module.exports.__wbg_subarray_270ff8dd5582c1ac = function(arg0, arg1, arg2) {
     const ret = getObject(arg0).subarray(arg1 >>> 0, arg2 >>> 0);
     return addHeapObject(ret);
 };
 
-export function __wbindgen_throw(arg0, arg1) {
+module.exports.__wbindgen_throw = function(arg0, arg1) {
     throw new Error(getStringFromWasm0(arg0, arg1));
 };
 
-export function __wbindgen_memory() {
+module.exports.__wbindgen_memory = function() {
     const ret = wasm.memory;
     return addHeapObject(ret);
 };
+
+const path = require('path').join(__dirname, 'cosmian_cover_crypt_bg.wasm');
+const bytes = require('fs').readFileSync(path);
+
+const wasmModule = new WebAssembly.Module(bytes);
+const wasmInstance = new WebAssembly.Instance(wasmModule, imports);
+wasm = wasmInstance.exports;
+module.exports.__wasm = wasm;
