+++ tests/wasm_lib/cosmian_cover_crypt/cosmian_cover_crypt_bg.wasm.d.ts
@@ -10,7 +10,7 @@ export function webassembly_decrypt_hybrid_header(a: number, b: number, c: numbe
 export function webassembly_encrypt_hybrid_block(a: number, b: number, c: number, d: number, e: number, f: number): void;
 export function webassembly_decrypt_hybrid_block(a: number, b: number, c: number, d: number, e: number, f: number): void;
 export function __wbindgen_add_to_stack_pointer(a: number): number;
-export function __wbindgen_malloc(a: number): number;
-export function __wbindgen_realloc(a: number, b: number, c: number): number;
-export function __wbindgen_free(a: number, b: number): void;
-export function __wbindgen_exn_store(a: number): void;
+export function __wbindgen_export_0(a: number): number;
+export function __wbindgen_export_1(a: number, b: number, c: number): number;
+export function __wbindgen_export_2(a: number, b: number): void;
+export function __wbindgen_export_3(a: number): void;
