+++ tests/wasm_lib/cosmian_cover_crypt/package.json
@@ -1,21 +1,21 @@
 {
-  "name": "cover_crypt",
+  "name": "cosmian_cover_crypt",
   "collaborators": [
     "Théophile Brezot <theophile.brezot@cosmian.com>",
     "Bruno Grieder <bruno.grieder@cosmian.com>"
   ],
   "description": "Key Policy attribute encryption based on subset cover",
-  "version": "6.0.1",
+  "version": "6.0.5",
   "license": "MIT/Apache-2.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/Cosmian/cover_crypt"
   },
   "files": [
-    "cover_crypt_bg.wasm",
-    "cover_crypt.js",
-    "cover_crypt.d.ts"
+    "cosmian_cover_crypt_bg.wasm",
+    "cosmian_cover_crypt.js",
+    "cosmian_cover_crypt.d.ts"
   ],
-  "main": "cover_crypt.js",
-  "types": "cover_crypt.d.ts"
+  "main": "cosmian_cover_crypt.js",
+  "types": "cosmian_cover_crypt.d.ts"
 }
