+++ tests/wasm_lib/cosmian_findex/cosmian_findex.d.ts
