+++ tests/wasm_lib/cosmian_findex/cosmian_findex_bg.wasm.d.ts
@@ -5,8 +5,8 @@ export function webassembly_upsert(a: number, b: number, c: number, d: number, e
 export function webassembly_graph_upsert(a: number, b: number, c: number, d: number, e: number, f: number, g: number, h: number): number;
 export function webassembly_search(a: number, b: number, c: number, d: number, e: number, f: number, g: number, h: number, i: number): number;
 export const __wbindgen_export_0: WebAssembly.Table;
-export function _dyn_core__ops__function__FnMut__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h38e12736b3baa3a1(a: number, b: number, c: number): void;
-export function __wbindgen_malloc(a: number): number;
-export function __wbindgen_realloc(a: number, b: number, c: number): number;
-export function __wbindgen_exn_store(a: number): void;
-export function wasm_bindgen__convert__closures__invoke2_mut__hc7e4f52cabe2049d(a: number, b: number, c: number, d: number): void;
+export function __wbindgen_export_1(a: number, b: number, c: number): void;
+export function __wbindgen_export_2(a: number): number;
+export function __wbindgen_export_3(a: number, b: number, c: number): number;
+export function __wbindgen_export_4(a: number): void;
+export function __wbindgen_export_5(a: number, b: number, c: number, d: number): void;
